moved nav tests back into single file
