+++ Source/Xamarin/Prism.Forms.Tests/Navigation/PageNavigationServiceFixture.cs
@@ -12,7 +12,6 @@ using Xunit;
 
 namespace Prism.Forms.Tests.Navigation
 {
-    [Collection("Navigation")]
     public class PageNavigationServiceFixture : IDisposable
     {
         PageNavigationContainerMock _container;
@@ -460,6 +459,370 @@ namespace Prism.Forms.Tests.Navigation
         }
 
         [Fact]
+        public async void Navigate_FromNavigationPage_ToContentPage_ByName()
+        {
+            var navigationService = new PageNavigationServiceMock(_container, _applicationProvider, _loggerFacade);
+            var rootPage = new Xamarin.Forms.NavigationPage();
+            ((IPageAware)navigationService).Page = rootPage;
+
+            await navigationService.NavigateAsync("ContentPage", useModalNavigation: false);
+
+            Assert.True(rootPage.Navigation.NavigationStack.Count == 1);
+            Assert.IsType<ContentPageMock>(rootPage.Navigation.NavigationStack[0]);
+        }
+
+        //TODO: rename tests to follow a new test naming convention 
+        [Fact]
+        public async Task Navigate_FromNavigationPage_WithoutChildPage_ToContentPage()
+        {
+            var recorder = new PageNavigationEventRecorder();
+            var navigationService = new PageNavigationServiceMock(_container, _applicationProvider, _loggerFacade, recorder);
+            var navigationPage = new NavigationPageEmptyMock(recorder);
+
+            ((IPageAware)navigationService).Page = navigationPage;
+            await navigationService.NavigateAsync("ContentPage");
+
+            Assert.Equal(0, navigationPage.Navigation.ModalStack.Count);
+            Assert.Equal(1, navigationPage.Navigation.NavigationStack.Count);
+            var contentPage = navigationPage.Navigation.NavigationStack.Last();
+            Assert.IsType<ContentPageMock>(contentPage);
+
+            var record = recorder.TakeFirst();
+            Assert.Equal(contentPage, record.Sender);
+            Assert.Equal(PageNavigationEvent.OnNavigatingTo, record.Event);
+
+            record = recorder.TakeFirst();
+            Assert.Equal(contentPage.BindingContext, record.Sender);
+            Assert.Equal(PageNavigationEvent.OnNavigatingTo, record.Event);
+
+            record = recorder.TakeFirst();
+            Assert.Equal(navigationPage, record.Sender);
+            Assert.Equal(PageNavigationEvent.OnNavigatedFrom, record.Event);
+
+            record = recorder.TakeFirst();
+            Assert.Equal(navigationPage.BindingContext, record.Sender);
+            Assert.Equal(PageNavigationEvent.OnNavigatedFrom, record.Event);
+
+            record = recorder.TakeFirst();
+            Assert.Equal(contentPage, record.Sender);
+            Assert.Equal(PageNavigationEvent.OnNavigatedTo, record.Event);
+
+            record = recorder.TakeFirst();
+            Assert.Equal(contentPage.BindingContext, record.Sender);
+            Assert.Equal(PageNavigationEvent.OnNavigatedTo, record.Event);
+
+            Assert.True(recorder.IsEmpty);
+        }
+
+        //TODO: reimplement test to check order of events when navigating in a navigationpage. because of reverse navigation, this no longer is valid.
+        [Fact]
+        public async Task NavigateAsync_From_ChildPageOfNavigationPage()
+        {
+            var recorder = new PageNavigationEventRecorder(); ;
+            var navigationService = new PageNavigationServiceMock(_container, _applicationProvider, _loggerFacade, recorder);
+            var contentPageMock = new ContentPageMock(recorder);
+            var navigationPage = new NavigationPageMock(recorder, contentPageMock);
+
+            // Navigate to Page2
+            ((IPageAware)navigationService).Page = contentPageMock;
+            await navigationService.NavigateAsync("SecondContentPageMock");
+
+            Assert.Equal(0, navigationPage.Navigation.ModalStack.Count);
+            Assert.Equal(2, navigationPage.Navigation.NavigationStack.Count);
+
+            var pageMock = navigationPage.Navigation.NavigationStack.Last();
+
+            Assert.IsType<SecondContentPageMock>(pageMock);
+
+            var record = recorder.TakeFirst();
+            Assert.Equal(pageMock, record.Sender);
+            Assert.Equal(PageNavigationEvent.OnNavigatingTo, record.Event);
+
+            record = recorder.TakeFirst();
+            Assert.Equal(pageMock.BindingContext, record.Sender);
+            Assert.Equal(PageNavigationEvent.OnNavigatingTo, record.Event);
+
+            record = recorder.TakeFirst();
+            Assert.Equal(contentPageMock, record.Sender);
+            Assert.Equal(PageNavigationEvent.OnNavigatedFrom, record.Event);
+
+            record = recorder.TakeFirst();
+            Assert.Equal(contentPageMock.BindingContext, record.Sender);
+            Assert.Equal(PageNavigationEvent.OnNavigatedFrom, record.Event);
+
+            record = recorder.TakeFirst();
+            Assert.Equal(pageMock, record.Sender);
+            Assert.Equal(PageNavigationEvent.OnNavigatedTo, record.Event);
+
+            record = recorder.TakeFirst();
+            Assert.Equal(pageMock.BindingContext, record.Sender);
+            Assert.Equal(PageNavigationEvent.OnNavigatedTo, record.Event);
+
+            Assert.True(recorder.IsEmpty);
+        }
+
+        //TODO: reimplement test to check order of events when navigating in a navigationpage. because of reverse navigation, this no longer is valid.
+        [Fact]
+        public async Task NavigateAsync_From_NavigationPage_With_ChildPage_And_DoesNotReplaseRootPage()
+        {
+            var recorder = new PageNavigationEventRecorder(); ;
+            var navigationService = new PageNavigationServiceMock(_container, _applicationProvider, _loggerFacade, recorder);
+            var contentPageMock = new ContentPageMock(recorder);
+            var contentPageMockViewModel = contentPageMock.BindingContext;
+            var navigationPage = new NavigationPageMock(recorder, contentPageMock);
+
+            // Navigate to Page2
+            ((IPageAware)navigationService).Page = contentPageMock;
+            await navigationService.NavigateAsync("SecondContentPageMock");
+
+            var secondContentPage = navigationPage.Navigation.NavigationStack.Last();
+            var secondContentPageViewModel = secondContentPage.BindingContext;
+
+            recorder.Clear();
+            // PopToRootAsync
+            ((IPageAware)navigationService).Page = navigationPage;
+            await navigationService.NavigateAsync("ContentPage");
+
+            Assert.Equal(0, navigationPage.Navigation.ModalStack.Count);
+            Assert.Equal(1, navigationPage.Navigation.NavigationStack.Count);
+
+            var rootPage = navigationPage.Navigation.NavigationStack.Last();
+            Assert.Equal(contentPageMock, rootPage);
+
+            var record = recorder.TakeFirst();
+            Assert.Equal(contentPageMock, record.Sender);
+            Assert.Equal(PageNavigationEvent.OnNavigatingTo, record.Event);
+
+            record = recorder.TakeFirst();
+            Assert.Equal(contentPageMockViewModel, record.Sender);
+            Assert.Equal(PageNavigationEvent.OnNavigatingTo, record.Event);
+
+            record = recorder.TakeFirst();
+            Assert.Equal(contentPageMock, record.Sender);
+            Assert.Equal(PageNavigationEvent.OnNavigatedFrom, record.Event);
+
+            record = recorder.TakeFirst();
+            Assert.Equal(contentPageMockViewModel, record.Sender);
+            Assert.Equal(PageNavigationEvent.OnNavigatedFrom, record.Event);
+
+            record = recorder.TakeFirst();
+            Assert.Equal(contentPageMock, record.Sender);
+            Assert.Equal(PageNavigationEvent.OnNavigatedTo, record.Event);
+
+            record = recorder.TakeFirst();
+            Assert.Equal(contentPageMock.BindingContext, record.Sender);
+            Assert.Equal(PageNavigationEvent.OnNavigatedTo, record.Event);
+
+            record = recorder.TakeFirst();
+            Assert.Equal(secondContentPage, record.Sender);
+            Assert.Equal(PageNavigationEvent.Destroy, record.Event);
+
+            record = recorder.TakeFirst();
+            Assert.Equal(secondContentPageViewModel, record.Sender);
+            Assert.Equal(PageNavigationEvent.Destroy, record.Event);
+
+            Assert.True(recorder.IsEmpty);
+        }
+
+        //TODO: reimplement test to check order of events when navigating in a navigationpage. because of reverse navigation, this no longer is valid.
+        //[Fact]
+        //public async Task NavigateAsync_From_NavigationPage_With_ChildPage_And_ReplaseRootPage()
+        //{
+        //    var recorder = new PageNavigationEventRecorder(); ;
+        //    var navigationService = new PageNavigationServiceMock(_container, _applicationProvider, _loggerFacade, recorder);
+        //    var secondContentPageMock = new SecondContentPageMock(recorder);
+        //    var secondContentPageMockViewModel = secondContentPageMock.BindingContext;
+        //    var navigationPage = new NavigationPageMock(recorder, secondContentPageMock);
+
+        //    // Navigate to Page2
+        //    ((IPageAware)navigationService).Page = secondContentPageMock;
+        //    await navigationService.NavigateAsync("MasterDetailPage");
+
+        //    var masterDetailPage = navigationPage.Navigation.NavigationStack.Last();
+        //    var masterDetailPageViewModel = masterDetailPage.BindingContext;
+
+        //    recorder.Clear();
+        //    // PopToRootAsync
+        //    ((IPageAware)navigationService).Page = navigationPage;
+        //    await navigationService.NavigateAsync("ContentPage");
+
+        //    Assert.Equal(0, navigationPage.Navigation.ModalStack.Count);
+        //    Assert.Equal(1, navigationPage.Navigation.NavigationStack.Count);
+
+        //    var contentPage = navigationPage.Navigation.NavigationStack.Last();
+        //    Assert.NotEqual(secondContentPageMock, contentPage);
+
+        //    var record = recorder.TakeFirst();
+        //    Assert.Equal(contentPage, record.Sender);
+        //    Assert.Equal(PageNavigationEvent.OnNavigatingTo, record.Event);
+
+        //    record = recorder.TakeFirst();
+        //    Assert.Equal(contentPage.BindingContext, record.Sender);
+        //    Assert.Equal(PageNavigationEvent.OnNavigatingTo, record.Event);
+
+        //    record = recorder.TakeFirst();
+        //    Assert.Equal(secondContentPageMock, record.Sender);
+        //    Assert.Equal(PageNavigationEvent.OnNavigatedFrom, record.Event);
+
+        //    record = recorder.TakeFirst();
+        //    Assert.Equal(secondContentPageMockViewModel, record.Sender);
+        //    Assert.Equal(PageNavigationEvent.OnNavigatedFrom, record.Event);
+
+        //    record = recorder.TakeFirst();
+        //    Assert.Equal(contentPage, record.Sender);
+        //    Assert.Equal(PageNavigationEvent.OnNavigatedTo, record.Event);
+
+        //    record = recorder.TakeFirst();
+        //    Assert.Equal(contentPage.BindingContext, record.Sender);
+        //    Assert.Equal(PageNavigationEvent.OnNavigatedTo, record.Event);
+
+        //    record = recorder.TakeFirst();
+        //    Assert.Equal(masterDetailPage, record.Sender);
+        //    Assert.Equal(PageNavigationEvent.Destroy, record.Event);
+
+        //    record = recorder.TakeFirst();
+        //    Assert.Equal(masterDetailPageViewModel, record.Sender);
+        //    Assert.Equal(PageNavigationEvent.Destroy, record.Event);
+
+        //    record = recorder.TakeFirst();
+        //    Assert.Equal(secondContentPageMock, record.Sender);
+        //    Assert.Equal(PageNavigationEvent.Destroy, record.Event);
+
+        //    record = recorder.TakeFirst();
+        //    Assert.Equal(secondContentPageMockViewModel, record.Sender);
+        //    Assert.Equal(PageNavigationEvent.Destroy, record.Event);
+
+        //    Assert.True(recorder.IsEmpty);
+        //}
+
+        //TODO: reimplement test to check order of events when navigating in a navigationpage. because of reverse navigation, this no longer is valid.
+        //[Fact]
+        //public async Task NavigateAsync_From_NavigationPage_When_NotClearNavigationStack()
+        //{
+        //    var recorder = new PageNavigationEventRecorder(); ;
+        //    var navigationService = new PageNavigationServiceMock(_container, _applicationProvider, _loggerFacade, recorder);
+        //    var contentPageMock = new ContentPageMock(recorder);
+        //    var navigationPage = new NavigationPageMock(recorder, contentPageMock);
+        //    navigationPage.ClearNavigationStackOnNavigation = false;
+
+        //    // Navigate to Page2
+        //    ((IPageAware)navigationService).Page = contentPageMock;
+        //    await navigationService.NavigateAsync("SecondContentPageMock");
+
+        //    var secondContentPage = navigationPage.Navigation.NavigationStack.Last();
+        //    var secondContentPageViewModel = secondContentPage.BindingContext;
+
+        //    recorder.Clear();
+        //    ((IPageAware)navigationService).Page = navigationPage;
+        //    await navigationService.NavigateAsync("ContentPage");
+
+        //    Assert.Equal(0, navigationPage.Navigation.ModalStack.Count);
+        //    Assert.Equal(3, navigationPage.Navigation.NavigationStack.Count);
+
+        //    var currentPage = navigationPage.Navigation.NavigationStack.Last();
+        //    Assert.NotEqual(contentPageMock, currentPage);
+
+        //    var record = recorder.TakeFirst();
+        //    Assert.Equal(currentPage, record.Sender);
+        //    Assert.Equal(PageNavigationEvent.OnNavigatingTo, record.Event);
+
+        //    record = recorder.TakeFirst();
+        //    Assert.Equal(currentPage.BindingContext, record.Sender);
+        //    Assert.Equal(PageNavigationEvent.OnNavigatingTo, record.Event);
+
+        //    record = recorder.TakeFirst();
+        //    Assert.Equal(secondContentPage, record.Sender);
+        //    Assert.Equal(PageNavigationEvent.OnNavigatedFrom, record.Event);
+
+        //    record = recorder.TakeFirst();
+        //    Assert.Equal(secondContentPageViewModel, record.Sender);
+        //    Assert.Equal(PageNavigationEvent.OnNavigatedFrom, record.Event);
+
+        //    record = recorder.TakeFirst();
+        //    Assert.Equal(currentPage, record.Sender);
+        //    Assert.Equal(PageNavigationEvent.OnNavigatedTo, record.Event);
+
+        //    record = recorder.TakeFirst();
+        //    Assert.Equal(currentPage.BindingContext, record.Sender);
+        //    Assert.Equal(PageNavigationEvent.OnNavigatedTo, record.Event);
+
+        //    Assert.True(recorder.IsEmpty);
+        //}
+
+        [Fact]
+        public async Task NavigateAsync_From_NavigationPage_When_NotClearNavigationStack_And_SamePage()
+        {
+            var recorder = new PageNavigationEventRecorder(); ;
+            var navigationService = new PageNavigationServiceMock(_container, _applicationProvider, _loggerFacade, recorder);
+            var contentPageMock = new ContentPageMock(recorder);
+            var navigationPage = new NavigationPageMock(recorder, contentPageMock);
+            navigationPage.ClearNavigationStackOnNavigation = false;
+
+            // Navigate to Page2
+            ((IPageAware)navigationService).Page = contentPageMock;
+            await navigationService.NavigateAsync("SecondContentPageMock");
+
+            var secondContentPage = navigationPage.Navigation.NavigationStack.Last();
+
+            recorder.Clear();
+            ((IPageAware)navigationService).Page = navigationPage;
+            await navigationService.NavigateAsync("SecondContentPageMock");
+
+            Assert.Equal(0, navigationPage.Navigation.ModalStack.Count);
+            Assert.Equal(2, navigationPage.Navigation.NavigationStack.Count);
+
+            var currentPage = navigationPage.Navigation.NavigationStack.Last();
+            Assert.Equal(secondContentPage, currentPage);
+
+            var record = recorder.TakeFirst();
+            Assert.Equal(secondContentPage, record.Sender);
+            Assert.Equal(PageNavigationEvent.OnNavigatingTo, record.Event);
+
+            record = recorder.TakeFirst();
+            Assert.Equal(secondContentPage.BindingContext, record.Sender);
+            Assert.Equal(PageNavigationEvent.OnNavigatingTo, record.Event);
+
+            record = recorder.TakeFirst();
+            Assert.Equal(secondContentPage, record.Sender);
+            Assert.Equal(PageNavigationEvent.OnNavigatedFrom, record.Event);
+
+            record = recorder.TakeFirst();
+            Assert.Equal(secondContentPage.BindingContext, record.Sender);
+            Assert.Equal(PageNavigationEvent.OnNavigatedFrom, record.Event);
+
+            record = recorder.TakeFirst();
+            Assert.Equal(secondContentPage, record.Sender);
+            Assert.Equal(PageNavigationEvent.OnNavigatedTo, record.Event);
+
+            record = recorder.TakeFirst();
+            Assert.Equal(secondContentPage.BindingContext, record.Sender);
+            Assert.Equal(PageNavigationEvent.OnNavigatedTo, record.Event);
+
+            Assert.True(recorder.IsEmpty);
+        }
+
+
+        [Fact]
+        public async Task DeepNavigate_ToNavigationPage_ToTabbedPage_ToContentPage()
+        {
+            var navigationService = new PageNavigationServiceMock(_container, _applicationProvider, _loggerFacade);
+            var rootPage = new Xamarin.Forms.ContentPage();
+            ((IPageAware)navigationService).Page = rootPage;
+
+            await navigationService.NavigateAsync("NavigationPage/ContentPage/TabbedPage/PageMock");
+
+            var navPage = rootPage.Navigation.ModalStack[0] as NavigationPageMock;
+            Assert.NotNull(navPage);
+
+            var contentPage = navPage.Navigation.NavigationStack[0] as ContentPageMock;
+            Assert.NotNull(contentPage);
+
+            var tabbedPage = navPage.Navigation.NavigationStack[1] as TabbedPageMock;
+            Assert.NotNull(tabbedPage.CurrentPage);
+            Assert.IsType<PageMock>(tabbedPage.CurrentPage);
+        }
+
+        [Fact]
         public async void DeepNavigate_From_ContentPage_To_ContentPage()
         {
             var navigationService = new PageNavigationServiceMock(_container, _applicationProvider, _loggerFacade);
@@ -662,6 +1025,21 @@ namespace Prism.Forms.Tests.Navigation
         }
 
         [Fact]
+        public async void DeepNavigate_ToTabbedPage_ToPage()
+        {
+            var navigationService = new PageNavigationServiceMock(_container, _applicationProvider, _loggerFacade);
+            var rootPage = new Xamarin.Forms.ContentPage();
+            ((IPageAware)navigationService).Page = rootPage;
+
+            await navigationService.NavigateAsync("TabbedPage/PageMock");
+
+            var tabbedPage = rootPage.Navigation.ModalStack[0] as TabbedPageMock;
+            Assert.NotNull(tabbedPage);
+            Assert.NotNull(tabbedPage.CurrentPage);
+            Assert.IsType<PageMock>(tabbedPage.CurrentPage);
+        }
+
+        [Fact]
         public async void DeepNavigate_ToCarouselPage_ToContentPage()
         {
             var navigationService = new PageNavigationServiceMock(_container, _applicationProvider, _loggerFacade);
@@ -675,6 +1053,299 @@ namespace Prism.Forms.Tests.Navigation
             Assert.NotNull(tabbedPage.CurrentPage);
             Assert.IsType<ContentPageMock>(tabbedPage.CurrentPage);
         }
+
+        #region MasterDetailPage
+
+        [Fact]
+        public async void Navigate_FromMasterDetailPage()
+        {
+            var navigationService = new PageNavigationServiceMock(_container, _applicationProvider, _loggerFacade);
+            var rootPage = new MasterDetailPageMock();
+            ((IPageAware)navigationService).Page = rootPage;
+
+            Assert.IsType<ContentPageMock>(rootPage.Detail);
+
+            await navigationService.NavigateAsync("TabbedPage");
+
+            Assert.IsType<TabbedPageMock>(rootPage.Detail);
+
+            await navigationService.NavigateAsync("CarouselPage");
+
+            Assert.IsType<CarouselPageMock>(rootPage.Detail);
+        }
+
+        [Fact]
+        public async void Navigate_FromMasterDetailPage_ToSamePage()
+        {
+            var navigationService = new PageNavigationServiceMock(_container, _applicationProvider, _loggerFacade);
+            var rootPage = new MasterDetailPageMock();
+            ((IPageAware)navigationService).Page = rootPage;
+
+            Assert.IsType<ContentPageMock>(rootPage.Detail);
+
+            await navigationService.NavigateAsync("TabbedPage");
+
+            var firstDetailPage = rootPage.Detail;
+
+            Assert.IsType<TabbedPageMock>(firstDetailPage);
+
+            await navigationService.NavigateAsync("TabbedPage");
+
+            Assert.Equal(firstDetailPage, rootPage.Detail);
+        }
+
+        [Fact]
+        public async void DeepNavigate_ToEmptyMasterDetailPage_ToContentPage()
+        {
+            var navigationService = new PageNavigationServiceMock(_container, _applicationProvider, _loggerFacade);
+            var rootPage = new Xamarin.Forms.ContentPage();
+            ((IPageAware)navigationService).Page = rootPage;
+
+            await navigationService.NavigateAsync("MasterDetailPage-Empty/ContentPage");
+
+            Assert.Equal(1, rootPage.Navigation.ModalStack.Count);
+            Assert.Equal(0, rootPage.Navigation.NavigationStack.Count);
+
+            var masterDetail = rootPage.Navigation.ModalStack[0] as MasterDetailPageEmptyMock;
+            Assert.NotNull(masterDetail);
+            Assert.NotNull(masterDetail.Detail);
+            Assert.IsType<ContentPageMock>(masterDetail.Detail);
+        }
+
+        [Fact]
+        public async void DeepNavigate_ToEmptyMasterDetailPage_ToContentPage_UseModalNavigation()
+        {
+            var navigationService = new PageNavigationServiceMock(_container, _applicationProvider, _loggerFacade);
+            var rootPage = new Xamarin.Forms.ContentPage();
+            ((IPageAware)navigationService).Page = rootPage;
+
+            await navigationService.NavigateAsync("MasterDetailPage-Empty/ContentPage", useModalNavigation: true);
+
+            Assert.Equal(1, rootPage.Navigation.ModalStack.Count);
+            Assert.Equal(0, rootPage.Navigation.NavigationStack.Count);
+            var masterDetail = rootPage.Navigation.ModalStack[0] as MasterDetailPageEmptyMock;
+            Assert.NotNull(masterDetail);
+            Assert.NotNull(masterDetail.Detail);
+            Assert.IsType<ContentPageMock>(masterDetail.Detail);
+        }
+
+        [Fact]
+        public async void DeepNavigate_ToEmptyMasterDetailPage_ToContentPage_NotUseModalNavigation()
+        {
+            var navigationService = new PageNavigationServiceMock(_container, _applicationProvider, _loggerFacade);
+            var rootPage = new ContentPageMock();
+            var navigationPage = new NavigationPage(rootPage);
+            ((IPageAware)navigationService).Page = rootPage;
+
+            Assert.Equal(1, rootPage.Navigation.NavigationStack.Count);
+            Assert.IsType<ContentPageMock>(navigationPage.CurrentPage);
+
+            await navigationService.NavigateAsync("MasterDetailPage-Empty/ContentPage", useModalNavigation: false);
+
+            Assert.Equal(0, rootPage.Navigation.ModalStack.Count);
+            Assert.Equal(2, rootPage.Navigation.NavigationStack.Count);
+            var masterDetail = rootPage.Navigation.NavigationStack[1] as MasterDetailPageEmptyMock;
+            Assert.NotNull(masterDetail);
+            Assert.NotNull(masterDetail.Detail);
+            Assert.IsType<ContentPageMock>(masterDetail.Detail);
+        }
+
+        [Fact]
+        public async void DeepNavigate_ToEmptyMasterDetailPage_ToNavigationPage()
+        {
+            var navigationService = new PageNavigationServiceMock(_container, _applicationProvider, _loggerFacade);
+            var rootPage = new Xamarin.Forms.ContentPage();
+            ((IPageAware)navigationService).Page = rootPage;
+
+            await navigationService.NavigateAsync("MasterDetailPage-Empty/NavigationPage");
+
+            var masterDetail = rootPage.Navigation.ModalStack[0] as MasterDetailPageEmptyMock;
+            Assert.NotNull(masterDetail);
+            Assert.NotNull(masterDetail.Detail);
+            Assert.IsType<NavigationPageMock>(masterDetail.Detail);
+        }
+
+        [Fact]
+        public async void DeepNavigate_ToEmptyMasterDetailPage_ToEmptyNavigationPage_ToContentPage()
+        {
+            var applicationProvider = new ApplicationProviderMock(null);
+            var navigationService = new PageNavigationServiceMock(_container, applicationProvider, _loggerFacade);
+
+            await navigationService.NavigateAsync("MasterDetailPage-Empty/NavigationPage-Empty/ContentPage");
+
+            var masterDetail = applicationProvider.MainPage as MasterDetailPageEmptyMock;
+            Assert.NotNull(masterDetail);
+            Assert.NotNull(masterDetail.Detail);
+            Assert.IsType<NavigationPageEmptyMock>(masterDetail.Detail);
+            Assert.Equal(0, masterDetail.Navigation.ModalStack.Count);
+            Assert.Equal(0, masterDetail.Navigation.NavigationStack.Count);
+            Assert.Equal(0, masterDetail.Detail.Navigation.ModalStack.Count);
+            Assert.Equal(1, masterDetail.Detail.Navigation.NavigationStack.Count);
+            Assert.IsType<ContentPageMock>(masterDetail.Detail.Navigation.NavigationStack.Last());
+        }
+
+        [Fact]
+        public async void DeepNavigate_ToEmptyMasterDetailPage_ToNavigationPage_ToContentPage()
+        {
+            var applicationProvider = new ApplicationProviderMock(null);
+            var navigationService = new PageNavigationServiceMock(_container, applicationProvider, _loggerFacade);
+
+            await navigationService.NavigateAsync("MasterDetailPage-Empty/NavigationPage/PageMock");
+
+            var masterDetail = applicationProvider.MainPage as MasterDetailPageEmptyMock;
+            Assert.NotNull(masterDetail);
+            Assert.NotNull(masterDetail.Detail);
+            Assert.IsType<NavigationPageMock>(masterDetail.Detail);
+            Assert.Equal(0, masterDetail.Navigation.ModalStack.Count);
+            Assert.Equal(0, masterDetail.Navigation.NavigationStack.Count);
+            Assert.Equal(0, masterDetail.Detail.Navigation.ModalStack.Count);
+            Assert.Equal(1, masterDetail.Detail.Navigation.NavigationStack.Count);
+            Assert.IsType<PageMock>(masterDetail.Detail.Navigation.NavigationStack.Last());
+        }
+
+        [Fact]
+        public async void DeepNavigate_ToMasterDetailPage_ToDifferentPage()
+        {
+            var navigationService = new PageNavigationServiceMock(_container, _applicationProvider, _loggerFacade);
+            var rootPage = new Xamarin.Forms.ContentPage();
+            ((IPageAware)navigationService).Page = rootPage;
+
+            await navigationService.NavigateAsync("MasterDetailPage/TabbedPage");
+
+            var masterDetail = rootPage.Navigation.ModalStack[0] as MasterDetailPageMock;
+            Assert.NotNull(masterDetail);
+            Assert.NotNull(masterDetail.Detail);
+            Assert.IsType<TabbedPageMock>(masterDetail.Detail);
+        }
+
+        [Fact]
+        public async void DeepNavigate_ToMasterDetailPage_ToSamePage_ToTabbedPage()
+        {
+            var navigationService = new PageNavigationServiceMock(_container, _applicationProvider, _loggerFacade);
+            var rootPage = new Xamarin.Forms.ContentPage();
+            ((IPageAware)navigationService).Page = rootPage;
+
+            await navigationService.NavigateAsync("MasterDetailPage/ContentPage/TabbedPage");
+
+            var masterDetail = rootPage.Navigation.ModalStack[0] as MasterDetailPageMock;
+            Assert.NotNull(masterDetail);
+            Assert.NotNull(masterDetail.Detail);
+            Assert.IsType<ContentPageMock>(masterDetail.Detail);
+
+            var tabbedPage = masterDetail.Navigation.ModalStack[0] as TabbedPageMock;
+            Assert.NotNull(tabbedPage);
+        }
+
+        [Fact]
+        public async void Navigate_FromMasterDetailPage_ToTabbedPage_IsPresented()
+        {
+            var navigationService = new PageNavigationServiceMock(_container, _applicationProvider, _loggerFacade);
+            var rootPage = new MasterDetailPageMock();
+            ((IPageAware)navigationService).Page = rootPage;
+            rootPage.IsPresentedAfterNavigation = true;
+
+            Assert.IsType<ContentPageMock>(rootPage.Detail);
+            Assert.False(rootPage.IsPresented);
+
+            await navigationService.NavigateAsync("TabbedPage");
+            Assert.IsType<TabbedPageMock>(rootPage.Detail);
+
+            Assert.True(rootPage.IsPresented);
+        }
+
+        [Fact]
+        public async void Navigate_FromMasterDetailPage_ToTabbedPage_IsNotPresented()
+        {
+            var navigationService = new PageNavigationServiceMock(_container, _applicationProvider, _loggerFacade);
+            var rootPage = new MasterDetailPageMock();
+            ((IPageAware)navigationService).Page = rootPage;
+            rootPage.IsPresentedAfterNavigation = false;
+
+            Assert.IsType<ContentPageMock>(rootPage.Detail);
+            Assert.False(rootPage.IsPresented);
+
+            await navigationService.NavigateAsync("TabbedPage");
+            Assert.IsType<TabbedPageMock>(rootPage.Detail);
+
+            Assert.False(rootPage.IsPresented);
+        }
+
+        [Fact]
+        public async void Navigate_FromMasterDetailPage_ToTabbedPage_IsPresented_FromViewModel()
+        {
+            var navigationService = new PageNavigationServiceMock(_container, _applicationProvider, _loggerFacade);
+            var rootPage = new MasterDetailPageEmptyMock();
+            ((IPageAware)navigationService).Page = rootPage;
+
+            ((MasterDetailPageEmptyMockViewModel)rootPage.BindingContext).IsPresentedAfterNavigation = true;
+
+            Assert.Null(rootPage.Detail);
+            Assert.False(rootPage.IsPresented);
+
+            await navigationService.NavigateAsync("TabbedPage");
+            Assert.IsType<TabbedPageMock>(rootPage.Detail);
+
+            Assert.True(rootPage.IsPresented);
+        }
+
+        [Fact]
+        public async void Navigate_FromMasterDetailPage_ToTabbedPage_IsNotPresented_FromViewModel()
+        {
+            var navigationService = new PageNavigationServiceMock(_container, _applicationProvider, _loggerFacade);
+            var rootPage = new MasterDetailPageEmptyMock();
+            ((IPageAware)navigationService).Page = rootPage;
+
+            ((MasterDetailPageEmptyMockViewModel)rootPage.BindingContext).IsPresentedAfterNavigation = false;
+
+            Assert.Null(rootPage.Detail);
+            Assert.False(rootPage.IsPresented);
+
+            await navigationService.NavigateAsync("TabbedPage");
+            Assert.IsType<TabbedPageMock>(rootPage.Detail);
+
+            Assert.False(rootPage.IsPresented);
+        }
+
+
+
+        [Fact]
+        public async void DeepNavigate_ToMasterDetailPage_ToNavigationPage_ToTabbedPage_ToPage()
+        {
+            var navigationService = new PageNavigationServiceMock(_container, _applicationProvider, _loggerFacade);
+            var rootPage = new Xamarin.Forms.ContentPage();
+            ((IPageAware)navigationService).Page = rootPage;
+
+            await navigationService.NavigateAsync("MasterDetailPage-Empty/NavigationPage/TabbedPage/PageMock");
+
+            var mdpPage = rootPage.Navigation.ModalStack[0] as MasterDetailPageEmptyMock;
+            var navPage = mdpPage.Detail as NavigationPageMock;
+            var tabbedPage = navPage.Navigation.NavigationStack[0] as TabbedPageMock;
+            Assert.NotNull(mdpPage);
+            Assert.NotNull(navPage);
+            Assert.NotNull(tabbedPage.CurrentPage);
+            Assert.IsType<PageMock>(tabbedPage.CurrentPage);
+        }
+
+        [Fact]
+        public async void DeepNavigate_ToMasterDetailPage_ToNavigationPage_ToContentPage_ToTabbedPage_ToPage()
+        {
+            var navigationService = new PageNavigationServiceMock(_container, _applicationProvider, _loggerFacade);
+            var rootPage = new Xamarin.Forms.ContentPage();
+            ((IPageAware)navigationService).Page = rootPage;
+
+            await navigationService.NavigateAsync("MasterDetailPage-Empty/NavigationPage/ContentPage/TabbedPage/PageMock");
+
+            var mdpPage = rootPage.Navigation.ModalStack[0] as MasterDetailPageEmptyMock;
+            var navPage = mdpPage.Detail as NavigationPageMock;
+            var contentPage = navPage.Navigation.NavigationStack[0] as ContentPageMock;
+            var tabbedPage = navPage.Navigation.NavigationStack[1] as TabbedPageMock;
+            Assert.NotNull(mdpPage);
+            Assert.NotNull(navPage);
+            Assert.NotNull(tabbedPage.CurrentPage);
+            Assert.IsType<PageMock>(tabbedPage.CurrentPage);
+        }
+
+        #endregion
 
         public void Dispose()
         {
